addded IP gen sources -- these can be user customized (and some are)
In particular, the DDR3 MIG IP has been modified. This commit
attempts to capture this, in case you have to re-customize the
IP for some reason. Probably the best way to go about it is to
make the IP from the .xci file (for your current version of PlanAhead)
and then run a diff from the file in the github repo so you can see where
modifications were made.